Force non-passive event for touch device
Touch event listeners had a breaking change in chrome 56. Passive events became the default in Chrome 56 and up. This PR's forces the listener to be non-passive, so that `preventDefault` is no longer being ignored. Without it, the page wants to navigate to the next / previous page when swiping on a touch screen.

See: https://developers.google.com/web/updates/2017/01/scrolling-intervention
Related: facebook/react/#8968